Warnings and GameModes
- Reworked code to get rid of warnings
- Moved NOSTR: support from AbstractType to CSFManager
- Some function renames
- Fixed a bug in BinaryReader, \0 is added for non-zero terminated
strings to straighten the buffer
- Fixed a bug in Log's toString method; it now returns a proper string
- Reworked MIXManager's map into a vector so it stores MIX files in the
right order
- LineSplitter's bool pop methods now support 'false', 'no' etc strings
instead of just a number
- Added Game Mode support

+++ Map_Editor_BaseCode/Map_Editor_BaseCode/Editor.Engine/Types/Waypoint.cpp
@@ -56,9 +56,9 @@ std::string Waypoint::getLetterIndex()
 	}
 	else
 	{
-		float divided = index / 26;
-		int letter1 = divided - 1;
-		int letter2 = index - 26 * divided;
+		float divided = (float) index / 26;
+		int letter1 = (int) divided - 1;
+		int letter2 = index - 26 * (int) divided;
 		char chX = 'A' + letter2;
 		char chY = 'A' + letter1;
